--- Morsemania/README.md
@@ -1,27 +1,0 @@
-# Prototipo de Morsemania
-
-## Requisitos
-
-- Python 3.10 o superior
-- Pygame
-
-## Instalación de dependencias
-
-Instala la librería pygame ejecutando:
-
-```
-pip install pygame
-```
-
-## Estructura del proyecto
-
-- `main.py`: archivo principal del juego
-- `config.py`: configuración de parámetros
-
-## Ejecución
-
-Para ejecutar el juego:
-
-```
-python main.py
-```
