--- Shooter/README.md
@@ -1,28 +1,0 @@
-# Prototipo de Shooter
-
-## Requisitos
-
-- Python 3.10 o superior
-- Pygame
-
-## Instalación de dependencias
-
-Instala la librería pygame ejecutando:
-
-```
-pip install pygame
-```
-
-## Estructura del proyecto
-
-- `main.py`: archivo principal del juego
-- `config.py`: configuración de parámetros
-- `Entities/`: contiene las clases principales (`Player`, `Projectile`, `Target`)
-
-## Ejecución
-
-Para ejecutar el juego:
-
-```
-python main.py
-```
