+++ src/config.py
@@ -1,0 +1,14 @@
+# config.py
+SCREEN_WIDTH = 900
+SCREEN_HEIGHT = 500
+FPS = 60
+GROUND_HEIGHT = 80
+GRAVITY = 2500.0
+MOVE_ACCEL = 8000.0
+MAX_SPEED = 380.0
+FRICTION = 6000.0
+JUMP_VELOCITY = 900.0
+BG_COLOR = (30, 30, 40)
+GROUND_COLOR = (40, 40, 50)
+PLAYER_COLOR = (220, 40, 40)
+TEXT_COLOR = (230, 230, 230)
